requirements.txt
python-binance==1.0.19
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0
tkcalendar
pyarrow 